refactor: speed up twitter DB accessing
