+++ packages/web3-providers/src/Twitter/apis/index.ts
@@ -1,3 +1,4 @@
+export * from './getDatabase.js'
 export * from './getSettings.js'
 export * from './getTokens.js'
 export * from './getUserNFTContainer.js'
