+++ packages/web3-providers/src/types/Twitter.ts
@@ -337,7 +337,6 @@ export namespace TwitterBaseAPI {
     export interface Provider {
         getSettings: () => Promise<Settings | undefined>
         getUserSettings: () => Promise<UserSettings>
-        staleUserSettings: () => Promise<void>
         /**
          * @param screenName
          * @param checkNFTAvatar With `checkNFTAvatar` true, will get user via web API directly,
